# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ExtractBody": {
        "additionalProperties": false,
        "properties": {
          "background_asset_id": {
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Background Asset Id",
            "type": "string"
          }
        },
        "required": [
          "background_asset_id"
        ],
        "title": "ExtractBody",
        "type": "object"
      },
      "ExtractedTheme": {
        "properties": {
          "message": {
            "default": "Colors extracted from background",
            "title": "Message",
            "type": "string"
          },
          "palette": {
            "items": {
              "type": "string"
            },
            "title": "Palette",
            "type": "array"
          },
          "preset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset"
          },
          "source": {
            "$ref": "#/components/schemas/ThemeSource"
          },
          "tokens": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Tokens",
            "type": "object"
          }
        },
        "required": [
          "tokens",
          "source",
          "preset",
          "palette"
        ],
        "title": "ExtractedTheme",
        "type": "object"
      },
...
      "FrameResponse": {
        "description": "A validated frame. The stored file itself is served by `/content`.",
        "properties": {
          "builtin": {
            "description": "Packaged with the app; can not be replaced or deleted",
            "title": "Builtin",
            "type": "boolean"
          },
          "bytes": {
            "title": "Bytes",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "family": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Built-in family id, e.g. midnight; null for uploads",
            "title": "Family"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "slot_transparency": {
            "items": {
              "type": "number"
            },
            "title": "Slot Transparency",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "template_key": {
            "title": "Template Key",
            "type": "string"
          },
          "template_version": {
            "title": "Template Version",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "used_by": {
            "description": "Event Profiles offering this frame to participants",
            "items": {
              "type": "string"
            },
            "title": "Used By",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "template_key",
          "template_version",
          "name",
          "status",
          "width",
          "height",
          "bytes",
          "sha256",
          "warnings",
          "slot_transparency",
          "created_at",
          "updated_at",
          "builtin",
          "family"
        ],
        "title": "FrameResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MainColoursBody": {
        "additionalProperties": false,
        "description": "The current complete theme and the two colours an organizer chose.",
        "properties": {
          "button": {
            "description": "Button colour (the main action colour)",
            "examples": [
              "#2F6FD6"
            ],
            "pattern": "^#[0-9A-Fa-f]{6}$",
            "title": "Button",
            "type": "string"
          },
          "text": {
            "description": "Text colour (headings and participant text)",
            "examples": [
              "#2F6FD6"
            ],
            "pattern": "^#[0-9A-Fa-f]{6}$",
            "title": "Text",
            "type": "string"
          },
          "tokens": {
            "maxProperties": 35,
            "minProperties": 35,
            "patternProperties": {
              "^[a-z_]{1,40}$": {
                "examples": [
                  "#2F6FD6"
                ],
                "pattern": "^#[0-9A-Fa-f]{6}$",
                "type": "string"
              }
            },
            "title": "Tokens",
            "type": "object"
          }
        },
        "required": [
          "tokens",
          "button",
          "text"
        ],
        "title": "MainColoursBody",
        "type": "object"
      },
      "MainColoursResponse": {
        "properties": {
          "button": {
            "description": "The Button colour as used (adjusted only for contrast)",
            "title": "Button",
            "type": "string"
          },
          "text": {
            "description": "The Text colour as used (adjusted only for contrast)",
            "title": "Text",
            "type": "string"
          },
          "tokens": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Every token, with the related ones regenerated",
            "title": "Tokens",
            "type": "object"
          }
        },
        "required": [
          "tokens",
          "button",
          "text"
        ],
        "title": "MainColoursResponse",
        "type": "object"
      },
...
      "RenameFrameBody": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "RenameFrameBody",
        "type": "object"
      },
...
      "ThemeSource": {
        "enum": [
          "preset",
          "extracted",
          "custom"
        ],
        "title": "ThemeSource",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Photobooth kiosk API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/frames": {
      "get": {
        "operationId": "list_frames_api_admin_frames_get",
        "parameters": [
          {
            "in": "query",
            "name": "template_key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[a-z0-9_]{1,64}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Template Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FrameResponse"
                  },
                  "title": "Response List Frames Api Admin Frames Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Frames",
        "tags": [
          "admin-frames"
        ]
      },
      "post": {
        "operationId": "upload_frame_api_admin_frames_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  },
                  "name": {
                    "examples": [
                      "Expo blue border"
                    ],
                    "type": "string"
                  },
                  "template_key": {
                    "examples": [
                      "strip_2x6"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "template_key",
                  "name",
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Upload Frame",
        "tags": [
          "admin-frames"
        ]
      }
    },
    "/api/admin/frames/{frame_id}": {
      "delete": {
        "operationId": "delete_frame_api_admin_frames__frame_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "frame_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Frame Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Frame",
        "tags": [
          "admin-frames"
        ]
      },
      "get": {
        "operationId": "frame_metadata_api_admin_frames__frame_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "frame_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Frame Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Frame Metadata",
        "tags": [
          "admin-frames"
        ]
      }
    },
    "/api/admin/frames/{frame_id}/content": {
      "get": {
        "description": "The original uploaded PNG, byte for byte.",
        "operationId": "frame_content_api_admin_frames__frame_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "frame_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Frame Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/png": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Frame Content",
        "tags": [
          "admin-frames"
        ]
      }
    },
    "/api/admin/frames/{frame_id}/name": {
      "put": {
        "operationId": "rename_frame_api_admin_frames__frame_id__name_put",
        "parameters": [
          {
            "in": "path",
            "name": "frame_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Frame Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameFrameBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Frame",
        "tags": [
          "admin-frames"
        ]
      }
    },
    "/api/admin/frames/{frame_id}/preview/{output_index}.jpg": {
      "get": {
        "description": "Sample output: numbered placeholder photos with this frame composited on top.",
        "operationId": "frame_preview_api_admin_frames__frame_id__preview__output_index__jpg_get",
        "parameters": [
          {
            "in": "path",
            "name": "frame_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Frame Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "output_index",
            "required": true,
            "schema": {
              "maximum": 16,
              "minimum": 1,
              "title": "Output Index",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Frame Preview",
        "tags": [
          "admin-frames"
        ]
      }
    },
    "/api/admin/frames/{frame_id}/replace": {
      "post": {
        "description": "Swap in a corrected PNG. Profiles keep their selection; the old file is kept unchanged.",
        "operationId": "replace_frame_api_admin_frames__frame_id__replace_post",
        "parameters": [
          {
            "in": "path",
            "name": "frame_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Frame Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Frame",
        "tags": [
          "admin-frames"
        ]
      }
    },
...
    "/api/admin/themes/extract": {
      "post": {
        "description": "A complete accessible theme from the dominant colours of an uploaded background image.\n\nRuns locally; the image is read, never changed. Nothing is saved until the profile is saved.",
        "operationId": "extract_theme_api_admin_themes_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractedTheme"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Extract Theme",
        "tags": [
          "admin-themes"
        ]
      }
    },
    "/api/admin/themes/main-colours": {
      "post": {
        "description": "Regenerate the related colours (shades, links, borders, focus, labels) from the Button\nand Text colours. Nothing is saved until the profile is saved.",
        "operationId": "theme_main_colours_api_admin_themes_main_colours_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MainColoursBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MainColoursResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Theme Main Colours",
        "tags": [
          "admin-themes"
        ]
      }
    },
...
  }
}